Finally it works! (for mouse-based input buttons 1,2,3)
Selections work between clicking the map AND clicking the dendrogram!

This was a weird one-line fix from the previous version so whatevs, you know?

Future upgrades: making the dictionary keys not based on button clicks, but, like, a palette of some kind (a la IDL cloudviz). It would be fairly straightforward to replace all the places where we ask for event.button with self.current_palette_choice (or something like that). We'd have to build a little UI that lets you play with the palette choice dudes, probably by clicking. Would we also want to build color selection into that? i hope not.